add library and team members edit, translation and delete

+++ src/main/java/org/hishatakaran/backend/repository/TeamMembersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamMembersRepository extends JpaRepository<TeamMembers, Integer> {
+public interface TeamMembersRepository extends JpaRepository<TeamMembers, Long> {
 }
